Update main.py
In this version of the code, the try block encompasses the entire program, including the definition and execution of the main() function. If an exception occurs anywhere in the program, it will be caught by the except block and the error message will be printed to the console.

Note that it's generally not a good idea to catch all exceptions like this because it can make it harder to debug errors in your code. It's better to catch only the specific exceptions that you expect to occur and handle them appropriately.